Move most main_window logic to separate files
