+++ main_window/main_window.py
@@ -1,19 +1,13 @@
 # This Python file uses the following encoding: utf-8
-import sys
-import pathlib
-import ipaddress
 from dataclasses import dataclass
 
-from PySide6.QtWidgets import QWidget, QFileDialog
-from PySide6.QtWebEngineWidgets import QWebEngineView
-from PySide6.QtCore import QTimer, QDateTime
+from PySide6.QtWidgets import QWidget
+from PySide6.QtCore import QTimer
 from PySide6.QtNetwork import QUdpSocket, QAbstractSocket, QNetworkInterface
 
 from pyqtgraph import mkPen, PlotDataItem
 from PySide6.QtGui import QPixmap
 import numpy as np
-
-import packet_spec
 
 # Important:
 # You need to run the following command to generate the ui_form.py file
@@ -46,8 +40,15 @@ class pid_window(QWidget):
 
 
 class MainWindow(QWidget):
+    # Imports for MainWindow functionality. Helps split large file into
+    # smaller modules containing related functionality
     from .udp import udp_connection_button_handler, join_multicast_group, \
         udp_receive_socket_data, udp_on_disconnected, udp_on_error
+    from .data_handlers import plot_point, filter_data, updateActState
+    from .recording_and_playback import recording_toggle_button_handler, \
+        open_file_button_handler, display_previous_data
+    from .logging import save_to_file
+
 
     def __init__(self, parent=None):
         super().__init__(parent)
@@ -163,75 +164,6 @@ class MainWindow(QWidget):
         self.ui.recordingToggleButton.toggled.connect(self.recording_toggle_button_handler)
         self.file_out = None
 
-    def plot_point(self, header, message):
-        plots = self.plots
-        match header.type:
-            case packet_spec.PacketType.CONTROL:
-                # Cannot reach since the we only receive telemetry data
-                pass
-            case packet_spec.PacketType.TELEMETRY:
-                match header.sub_type:
-                    case packet_spec.TelemetryPacketSubType.TEMPERATURE:
-                        temperatureId:str = "t" + str(message.id)
-                        plots[temperatureId].points = np.append(plots[temperatureId].points, np.array([[message.time_since_power, message.temperature]]), axis=0)
-                        plots[temperatureId].data_line.setData(plots[temperatureId].points)
-                    case packet_spec.TelemetryPacketSubType.PRESSURE:
-                        pressureId:str = "p" + str(message.id)
-                        plots[pressureId].points = np.append(plots[pressureId].points, np.array([[message.time_since_power, message.pressure]]), axis=0)
-                        plots[pressureId].data_line.setData(plots[pressureId].points)
-                    case packet_spec.TelemetryPacketSubType.MASS:
-                        tankMass:str = "tank_mass"
-                        plots[tankMass].points = np.append(plots[tankMass].points, np.array([[message.time_since_power, message.mass]]), axis=0)
-                        plots[tankMass].data_line.setData(plots[tankMass].points)
-                    case packet_spec.TelemetryPacketSubType.ARMING_STATE:
-                        pass
-                    case packet_spec.TelemetryPacketSubType.ACT_STATE:
-                        pass
-                    case packet_spec.TelemetryPacketSubType.WARNING:
-                        pass
-
-    def recording_toggle_button_handler(self):
-        pathlib.Path('recording').mkdir(parents=True, exist_ok=True)
-        if self.ui.recordingToggleButton.isChecked() == True:
-            file_name = './recording/'
-            file_name += QDateTime.currentDateTime().toString("yyyy-MM-dd_HH-mm")
-            file_name += '.dump'
-            self.file_out = open(file_name, "a+b")
-        else:
-            self.file_out.close()
-
-    def display_previous_data(self,data):
-            ptr = 0
-            data_len = len(data)
-            while(ptr < data_len):
-                header = data[ptr:ptr + 2]
-                ptr += 2
-                data_header = packet_spec.parse_packet_header(header)
-                message_bytes_length = packet_spec.packet_message_bytes_length(data_header)
-                message = data[ptr:ptr + message_bytes_length]
-                data_message = packet_spec.parse_packet_message(data_header, message)
-                ptr += message_bytes_length
-                self.plot_point(data_header, data_message)
-
-    def open_file_button_handler(self):
-            file_path, _ = QFileDialog.getOpenFileName(self, "Open Previous File", "recording", "Dump file(*.dump);;All files (*)")
-
-            # If a file is selected, read its contents
-            if file_path:
-                self.ui.logOutput.append(f"Reading data from {file_path}")
-                with open(file_path, 'rb') as file:
-                    data = file.read()
-                    self.display_previous_data(data)
-                self.ui.logOutput.append("Data loaded")
-                    
-    def filter_data(self):
-        for key in self.plots:
-            if self.plots[key].points.size == 0:
-                continue
-            min_time: int = self.plots[key].points[:,0].max() - self.time_range
-            self.plots[key].points = self.plots[key].points[self.plots[key].points[:,0] >= min_time]
-            self.plots[key].data_line.setData(self.plots[key].points)
-
     # Handles when the window is closed, have to make sure to disconnect the TCP socket
     def closeEvent(self, event):
         if self.padUDPSocket.state() == QAbstractSocket.SocketState.ConnectedState:
@@ -239,128 +171,7 @@ class MainWindow(QWidget):
             self.padUDPSocket.waitForDisconnected()
 
         event.accept()
-
-    #UI updater, when actuator state messages are sent they will be updated with this function
-    def updateActState(self, message):
-        # Huge match case to match each valve id to a label in the UI
-        match message.id:
-            case 0:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.cv1State.setText("CLOSED")
-                    self.ui.cv1State.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.cv1State_tabpid.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.cv1State_tabpid.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.cv1State.setText("OPEN")
-                    self.ui.cv1State.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.cv1State_tabpid.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.cv1State_tabpid.setText("OPEN")
-            case 1:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv1State.setText("CLOSED")
-                    self.ui.xv1State.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv1State_tabpid.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv1State_tabpid.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv1State.setText("OPEN")
-                    self.ui.xv1State.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv1State_tabpid.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv1State_tabpid.setText("OPEN")
-            case 2:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv2State.setText("CLOSED")
-                    self.ui.xv2State.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv2State_tabpid.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv2State_tabpid.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv2State.setText("OPEN")
-                    self.ui.xv2State.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv2State_tabpid.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv2State_tabpid.setText("OPEN")
-            case 3:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv3State.setText("CLOSED")
-                    self.ui.xv3State.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv3State_tabpid.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv3State_tabpid.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv3State.setText("OPEN")
-                    self.ui.xv3State.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv3State_tabpid.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv3State_tabpid.setText("OPEN")
-            case 4:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv4State.setText("CLOSED")
-                    self.ui.xv4State.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv4State_tabpid.setStyleSheet("background-color: rgb(255, 80, 80)")
-                    self.ui.xv4State_tabpid.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv4State.setText("OPEN")
-                    self.ui.xv4State.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv4State_tabpid.setStyleSheet("background-color: rgb(80, 255, 80)")
-                    self.ui.xv4State_tabpid.setText("OPEN")
-            case 5:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv5State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv5State.setText("OPEN")
-            case 6:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv6State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv6State.setText("OPEN")
-            case 7:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv7State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv7State.setText("OPEN")
-            case 8:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv8State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv8State.setText("OPEN")
-            case 9:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv9State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv9State.setText("OPEN")
-            case 10:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv10State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv10State.setText("OPEN")
-            case 11:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv11State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv11State.setText("OPEN")
-            case 12:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.xv12State.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.xv12State.setText("OPEN")
-            case 13:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.quickDisconnectState.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.quickDisconnectState.setText("OPEN")
-            case 14:
-                if(message.state == packet_spec.ActuatorState.OFF):
-                    self.ui.igniterState.setText("CLOSED")
-                elif(message.state == packet_spec.ActuatorState.ON):
-                    self.ui.igniterState.setText("OPEN")
     
     def show_new_window(self, checked):
         self.w = pid_window()
         self.w.show()
-
-    #Creates a file or overwrites existing one, and writes the text in the logOutput into the file
-    def save_to_file(self):
-        pathlib.Path('recording').mkdir(parents=True, exist_ok=True)
-        file_name = './recording/'
-        file_name += QDateTime.currentDateTime().toString("yyyy-MM-dd_HH-mm")
-        file_name += '.dump'
-        f = open(file_name, "w")
-        f.write(self.ui.logOutput.toPlainText())
-        f.close()
-        self.ui.logOutput.append(f"Exported logs to {file_name}")
