--- main_window/plots.py
@@ -1,4 +1,0 @@
-"""plots.py
-
-Contains all functions related to plotting points on graph
-"""
